pyproject.toml
[tool.poetry]
name = "code-interpreter"
version = "0.1.0"
description = ""
authors = ["RutamBhagat <rutambhagat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.16"
black = "^24.4.2"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.4"
langchain-experimental = "^0.0.57"
qrcode = "^7.4.2"
pandas = "^2.2.2"
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
